remove StatusBar2 and copy StatasBar2 to StatusBar
