+++ components/pool/CollateralRatioBar.tsx
@@ -11,8 +11,6 @@ import { SimpleTooltip } from "components/shared/SimpleTooltip";
 import { PoolConfig } from "interfaces/pool";
 import { useAppData } from "context/AppDataContext";
 import StatusBar from "components/pool/StatusBar";
-import MyProgressBar from "components/pool/StatusBar2";
-import OuterProgressBar from "components/pool/StatusBar2";
 
 const CollateralRatioBar = ({ poolData }: { poolData?: PoolConfig }) => {
   const { t } = useTranslation();
@@ -64,10 +62,7 @@ const CollateralRatioBar = ({ poolData }: { poolData?: PoolConfig }) => {
                 borderRadius="10px"
                 value={liquidationPercentage}
               /> */}
-              {/* <StatusBar progress={50} /> */}
-              {/* <MyProgressBar success={70} warning={20} danger={10} striped animated={false}/> */}
-              {/* <OuterProgressBar success={70} warning={20} danger={10} striped={true} animated={false} /> */}
-              <MyProgressBar
+              <StatusBar
                 success={65}
                 warning={20}
                 danger={15}
